Moved Login to LoginUser, adding context soon
