+++ src/screens/HonorCodeScreen.js
@@ -1,46 +1,9 @@
 import React from 'react';
 import {View, StyleSheet, Text, Button} from 'react-native'
 import Hyperlink from 'react-native-hyperlink'
-import * as AuthSession from 'expo-auth-session'
-
+import { LoginUser } from '../providers/LoginUser'
 
 export const HonorCodeScreen = ({route, navigation})=>{
-
-    const signIn = async () => {
-        console.log("weee")
-        const infoResponse = await fetch('https://tutor.jakegut.com/auth/code/url')
-        const info = await infoResponse.json()
-        const redirectUrl = AuthSession.makeRedirectUri({useProxy: true});
-        console.log(redirectUrl);
-        let result = await AuthSession.startAsync({
-            authUrl:
-              `https://accounts.google.com/o/oauth2/auth?` +
-              `&client_id=${info['client_id']}` +
-              `&redirect_uri=${encodeURIComponent(redirectUrl)}` +
-              `&response_type=code` +
-              `&access_type=offline` +
-              `&scope=${encodeURIComponent(info['scopes'].join(' '))}` + 
-              `&prompt=consent`,
-          });
-        console.log(result);
-        if (result['errorCode'] == null) {
-            console.log("Getting accesstoken...")
-            fetch("https://tutor.jakegut.com/auth/swap", {
-                method: 'POST',
-                headers:{
-                    "Content-Type": "application/json"
-                },
-                body: JSON.stringify({"code": result['params']['code'], "redirect_uri": redirectUrl})
-            })
-            .then(response => response.json())
-            .then((data) => {
-                console.log(data)
-                navigation.navigate('HomeTabs')
-
-            })
-            .catch(err => console.log(err))
-        }
-    }
 
     return (
         <View style={styles.container}>
@@ -70,7 +33,7 @@ export const HonorCodeScreen = ({route, navigation})=>{
                         style={styles.decisionBox}
                     >
                         <Button
-                        onPress={()=>signIn()}
+                        onPress={()=>LoginUser(navigation)}
                         title="Yes"
                         />
                         <Button
